Normative: End `this` TDZ before running initializers
Previously, the TDZ of `this` in a subclass constructor ended when
the `super()` call was entirely finished. However, `this` is visible
as not being in TDZ running initializers. For example:

{
  let f;
  class A extends (class {}) {
    x = (this, f());
    constructor() { f = () => this; super(); }
  };
  new A;
}

Previously, `f()` would lead to a ReferenceError in this case, while
`this` would evaluate to the new instance under construction.

This patch tweaks the TDZ to end a little bit sooner, so that `this`
is defined inside of the constructor's scope *before* the initializers
are run. In the above example, there would be no ReferenceError
under the semantics in this patch.

Thanks to Caitlin Potter, Kevin Gibbons and Jordan Harband for
contributing to the discussion that led to this change.